Extract functions to another file
